provider and database initialization factored out from app.js to DbContext.js
